Update test name, move RTypeGenCoverageChecker to new file
